handle all unhappy scenarios
I have used mirage to mock out the interaction with the file url, this
allows me to simulate error scenarios in the tests.

I made the service return flash messages if there are any errors.
This makes the service responsible for dealing with anything regarding
fetching and parsing of the customer list allowing these steps to be
tested fully in one place.