implementation of distance service happy path
I read the wikipedia article on Great-circle distance and noticed it
stated that the haversine formula is better over small distances. Having
read this i looked for an existing implementation of the haversine
formula, i hope this doesn't mean i have skipped an important part of
the exercise! The notes i have been given do also state that the code
should be submitted as in a production shippable state. It was with this
in mind that i chose an existing well maintained implementation of the
formula. This would always be preference where existing packages meet
the requirements because of the time savings of not reinventing the wheel
and having had a larger audience to contribute to the solution.

I selected the package with the most stars and made sure that it had
multiple committers and the fact that it has been around since 2013 and
has been committed to recently made me more confident.

I have included a couple of sanity checks in the test just to make extra
sure the implementation is sound.